+++ index.js
@@ -7,35 +7,6 @@ const socketIO = require("socket.io");
 const {instrument} = require( "@socket.io/admin-ui"); 
 const PORT = process.env.PORT || 3000;
 
-//Connect to MongoDB
-const dbURI = 'mongodb+srv://jkmoore:sypzeg-Mupxit-2zudba@cluster0.bfd5u.mongodb.net/myFirstDatabase?retryWrites=true&w=majority';
-const mongoose = require('mongoose');
-const Schema = mongoose.Schema;
-mongoose.connect(dbURI, { useNewUrlParser: true , useUnifiedTopology: true });
-mongoose.connection.once('open', function(){
-  console.log('Connected to MongoDB');
-}).on('error', function(error){
-  console.log('error is:', error);
-});
-
-//setup requirejs to require schema module on frontend (submit.js)
-var requirejs = require("requirejs");
-requirejs.config({
-  nodeRequire: require,
-});
-/*
-//Create message schema
-const messageSchema = new Schema({
-  message: String
-})
-const Message = mongoose.model('Message', messageSchema);
-module.exports = Message;
-*/ 
-/* How to create and save a message
-const testMessage = new Message({
-  message: 'testing message with mongo'
-});
-testMessage.save(); */
 
 const INDEX = "/index.html";
 const DASHBOARD = "public/Dashboard.html";
@@ -72,6 +43,16 @@ instrument(io, {
   },
 });
 
+ //setup requirejs 
+  var requirejs = require("requirejs");
+  requirejs.config({
+    nodeRequire: require,
+  });
+  
+  //calls a module that connects to mongodb 
+  var mongoConnect = require('./public/js/mongoModule.js');
+  mongoConnect.db();
+
 app.get("/", (req, res) => {
   res.sendFile(INDEX, { root: __dirname });
 });
@@ -99,7 +80,7 @@ app.get("/about", (req, res) => {
 app.get("/message", (req, res) => {
   res.sendFile(MESSAGE, { root: __dirname });
 });
-
+/*
 const bodyParser = require('body-parser');
 app.use(bodyParser.urlencoded({extended: true}))
 app.post("/submit", (req, res) => {
@@ -107,7 +88,7 @@ app.post("/submit", (req, res) => {
   console.log(req.body.feedback);
   res.redirect("/feedbackDone.html");
 })
-
+*/
 io.on("connection", (socket) => {
   console.log("Client connected via socket (not in a room yet)");
